+++ Sources/Puppy/FileRotationLogger.swift
@@ -1,97 +1,21 @@
 import Foundation
 
-public class FileRotationLogger: BaseLogger {
+public class FileRotationLogger: FileLogger {
 
     public typealias ByteCount = UInt64
     public var maxFileSize: ByteCount = 10 * 1024 * 1024
     public var maxArchivedFilesCount: UInt8 = 5
 
-    private var fileHandle: FileHandle!
-    private let fileURL: URL
-
     public weak var delegate: FileRotationLoggerDeletate?
 
     public init(_ label: String, fileURL: URL) throws {
-        self.fileURL = fileURL
-        debug("fileURL is \(fileURL).")
-        super.init(label)
-        try validateFileURL(fileURL)
-        try openFile()
-    }
-
-    deinit {
-        closeFile()
+        try super.init(label, fileURL: fileURL)
     }
 
     public override func log(_ level: LogLevel, string: String) {
         rotateFiles()
-
-        do {
-            _ = try fileHandle?.seekToEndCompatible()
-            if let data = (string + "\r\n").data(using: .utf8) {
-                // swiftlint:disable force_try
-                try! fileHandle?.writeCompatible(contentsOf: data)
-                // swiftlint:enable force_try
-            }
-        } catch {
-            print("seekToEnd error. error is \(error.localizedDescription).")
-        }
-
+        super.log(level, string: string)
         rotateFiles()
-    }
-
-    public func delete(_ url: URL) throws {
-        do {
-            try queue!.sync {
-                try FileManager.default.removeItem(at: url)
-            }
-        } catch {
-            throw FileError.deletingFailed(at: url)
-        }
-    }
-
-    private func validateFileURL(_ url: URL) throws {
-        if url.hasDirectoryPath {
-            throw FileError.isNotFile(url: url)
-        }
-    }
-
-    private func openFile() throws {
-        closeFile()
-        let directoryURL = fileURL.deletingLastPathComponent()
-        do {
-            try FileManager.default.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
-            debug("created directoryURL is \(directoryURL).")
-        } catch {
-            throw FileError.creatingDirectoryFailed(at: directoryURL)
-        }
-
-        if !FileManager.default.fileExists(atPath: fileURL.path) {
-            let successful = FileManager.default.createFile(atPath: fileURL.path, contents: nil, attributes: nil)
-            if successful {
-                debug("succeeded in creating filePath.")
-            } else {
-                throw FileError.creatingFileFailed(at: fileURL)
-            }
-        } else {
-            debug("filePath exists. filePath is \(fileURL.path).")
-        }
-
-        if fileHandle == nil {
-            do {
-                fileHandle = try FileHandle(forWritingTo: fileURL)
-            } catch {
-                throw FileError.writingFailed(at: fileURL)
-            }
-        }
-    }
-
-    private func closeFile() {
-        if fileHandle != nil {
-            fileHandle.synchronizeFile()
-            fileHandle.closeFile()
-            fileHandle = nil
-        }
     }
 
     private func rotateFiles() {